feat: add screen() API for QA Wolf flows and mount TM v2 on hub server
QA Wolf flows bind OCR screens via screen(name) after configure({ page, storage }) — no dist imports or createOcrScreen(page, name). Mount GCS template manager at /template-manager on port 2020, add generated.ts catalog, and wire push/pull for team-storage screens.
