+++ packages/core/src/author/index.ts
@@ -4,7 +4,7 @@ export type { DetectScreenResult, BoxesFile } from './detect.js';
 export { applyScreen } from './apply.js';
 export type { FirstPass, FirstPassElement, FirstPassPart, FirstPassSection, ApplyScreenResult, AppliedElement, AppliedSection } from './apply.js';
 
-export { writeScreenCatalog, readScreenCatalog, screenCatalogSource } from './catalog.js';
+export { writeScreenCatalog, readScreenCatalog, screenCatalogSource, screenCatalogPath } from './catalog.js';
 export type { ScreenCatalog } from './catalog.js';
 
 export { showAnnotated } from './show.js';
