+++ packages/tools/index.html
@@ -160,8 +160,12 @@
 
     <h2>Tools</h2>
     <div class="grid">
+      <a class="card" href="/template-manager">
+        <strong>Template Manager (GCS)</strong>
+        <span>Author screens, pull/push team-storage, detect boxes, export <code>index.json</code>.</span>
+      </a>
       <a class="card" href="/template-manager.html">
-        <strong>Template Manager</strong>
+        <strong>Template Manager (legacy)</strong>
         <span>Crop fields, export <code>config.ts</code>, test OCR on the diff.</span>
       </a>
     </div>
